Created Visualizations, Cleaned up Data-generation, cleaned up some of the meshes

+++ generate_fracture.py
@@ -1,6 +1,9 @@
+import shutil
+
 import numpy as np
 import igl
 import tetgen
+import open3d as o3d
 
 from fracture_modes import fracture_utility as fracture
 # import fracture_utility as fracture
@@ -94,7 +97,7 @@ def generate_fracture(v_fine, f_fine, num_faces=100, output_folder="./results"):
     return modes
 
 
-def create_modes(v_fine, f_fine, num_faces=400, output_folder="./results"):
+def create_modes(v_fine, f_fine, num_faces=4000, output_folder="./results"):
     v_fine = np.asarray(v_fine)
     f_fine = np.asarray(f_fine)
     print("v_fine", len(v_fine))
@@ -138,7 +141,8 @@ def create_modes(v_fine, f_fine, num_faces=400, output_folder="./results"):
     return modes, v, f
 
 
-def generate_multiple_fractures(modes, num_impacts, v, f, category_name, dataset_name, pcd, mesh, volume_constraint=(1 / 50)):
+def generate_multiple_fractures(modes, num_impacts, v, f, category_name, dataset_name, pcd, mesh, ply_file,
+                                volume_constraint=(1 / 50)):
     B, FI = igl.random_points_on_mesh(1000 * num_impacts, v, f)
     FI[FI >= len(f)] = len(f) - 1  # fixes bug inside pyigl library which also includes indices equal to len(faces)
     B = np.vstack((B[:, 0], B[:, 0], B[:, 0], B[:, 1], B[:, 1], B[:, 1], B[:, 2], B[:, 2], B[:, 2])).T
@@ -153,12 +157,12 @@ def generate_multiple_fractures(modes, num_impacts, v, f, category_name, dataset
     running_num = 0
     num_tries = 0
     for i in range(P.shape[0]):
-        num_tries+=1
+        num_tries += 1
         contact_point = P[i, :]
-        direction = np.mean(modes.fine_vertices)-contact_point
+        direction = mesh.triangle_normals[FI[i]]
         # direction =  -np.copy(contact_point) / np.linalg.norm(np.copy(contact_point))
         modes.impact_projection(contact_point=contact_point, direction=direction, threshold=sigmas[i],
-                                num_modes_used=30)
+                                num_modes_used=20)
         min_volume = volume_constraint * total_vol / (modes.n_pieces_after_impact)
         current_min_volume = total_vol
         for j in range(modes.n_pieces_after_impact):
@@ -171,7 +175,7 @@ def generate_multiple_fractures(modes, num_impacts, v, f, category_name, dataset
         # print(modes.n_pieces_after_impact > 1, modes.n_pieces_after_impact < 100, new, valid_volume)
         if 1 < modes.n_pieces_after_impact < 100 and new and valid_volume:
             write_to_file(modes, output_folder=category_name, dataset_name=dataset_name, index=running_num, pcd=pcd,
-                          contact_point=contact_point, direction=direction, mesh=mesh)
+                          contact_point=contact_point, direction=direction, mesh=mesh, ply_file=ply_file)
             running_num += 1
             print("running_num: ", running_num)
             print("after this amount of tries: ", num_tries)
@@ -189,70 +193,64 @@ def generate_multiple_fractures(modes, num_impacts, v, f, category_name, dataset
             break
 
 
-
-def write_to_file(modes, output_folder, dataset_name, index, pcd, contact_point, direction, mesh):
-    if not os.path.exists(dataset_name):
-        os.makedirs(dataset_name)
-    # if not os.path.exists(os.path.join(dataset_name, output_folder)):
-    os.makedirs(os.path.join(dataset_name, output_folder), exist_ok=True)
-    os.makedirs(os.path.join(dataset_name, output_folder, "points"), exist_ok=True)
-    os.makedirs(os.path.join(dataset_name, output_folder, "points_label"), exist_ok=True)
-    os.makedirs(os.path.join(dataset_name, output_folder, "impulse_info"), exist_ok=True)
-    os.makedirs(os.path.join(dataset_name, output_folder, "mesh"), exist_ok=True)
-    os.makedirs(os.path.join(dataset_name, output_folder, "scaled_mesh"), exist_ok=True)
-
-    #retrieve pointcloud
-    pointcloud = pcd
-
-    # write pointcloud to .ply file
-    pcd_path = os.path.join(dataset_name, output_folder, "points", str(index) + ".pcd")
+def write_pcd_to_file(dataset_name, output_folder, pcd):
+    # write pointcloud to .pcd file
+    pcd_path = os.path.join(dataset_name, output_folder, "pointcloud.pcd")
+    if os.path.exists(pcd_path):
+        return
     with open(pcd_path, "w") as f:
         for point in np.asarray(pcd.points):
             f.write(f"{point[0]} {point[1]} {point[2]}\n")
 
-    #retrieve segmentation
+
+def write_segmentation_to_file(pointcloud, modes, mesh, dataset_name, output_folder, index):
+    os.makedirs(os.path.join(dataset_name, output_folder, "points_label"), exist_ok=True)
+    # retrieve segmentation
     _, labels, scaled_vertices = find_scipy_mapping(pointcloud, modes, mesh)
 
-    #write segmentation to .seg file
+    # write segmentation to .seg file
     seg_path = os.path.join(dataset_name, output_folder, "points_label", str(index) + ".seg")
     with open(seg_path, "w") as f:
         for label in labels:
             f.write(f"{label}\n")
 
-    #retrieve impulse info
 
+def write_impulse_to_file(contact_point, direction, dataset_name, output_folder, index):
+    os.makedirs(os.path.join(dataset_name, output_folder, "impulse_info"), exist_ok=True)
+
+    # retrieve impulse info
     impulse_info = np.concatenate((contact_point, direction))
 
-    #write impulse info to file or append it to pointcloud data
+    # write impulse info to file or append it to pointcloud data
     impulse_path = os.path.join(dataset_name, output_folder, "impulse_info", str(index) + ".imp")
-    pcd_min = np.min(pcd.points, axis=0)
-    pcd_max = np.max(pcd.points, axis=0)
-    mesh_min = np.min(modes.fine_vertices, axis=0)
-    mesh_max = np.max(modes.fine_vertices, axis=0)
-    contact_point = (contact_point - mesh_min) / (mesh_max - mesh_min) * (pcd_max - pcd_min) + pcd_min
 
     with open(impulse_path, "w") as f:
-        # print(impulse_info)
-        f.write(f"{contact_point[0]} {contact_point[1]} {contact_point[2]} {impulse_info[3]} {impulse_info[4]} {impulse_info[5]}\n")
-
-    mesh_path = os.path.join(dataset_name, output_folder, "mesh", str(index) + ".mesh")
-    with open(mesh_path, "w") as f:
-        for vertex in mesh.vertices:
-            f.write(f"{vertex[0]} {vertex[1]} {vertex[2]}\n")
-
-        f.write("\n")
-
-        for face in mesh.triangles:
-            f.write(f"{face[0]} {face[1]} {face[2]}\n")
+        f.write(
+            f"{contact_point[0]} {contact_point[1]} {contact_point[2]} {impulse_info[3]} {impulse_info[4]} {impulse_info[5]}\n")
 
 
-    scaled_mesh_path = os.path.join(dataset_name, output_folder, "scaled_mesh", str(index) + ".mesh")
-    with open(scaled_mesh_path, "w") as f:
-        for vertex in scaled_vertices:
-            f.write(f"{vertex[0]} {vertex[1]} {vertex[2]}\n")
+def write_mesh_to_file(dataset_name, output_folder, index, mesh):
+    os.makedirs(os.path.join(dataset_name, output_folder, "mesh"), exist_ok=True)
+    mesh_path = os.path.join(dataset_name, output_folder, "mesh", str(index) + ".obj")
+    o3d.io.write_triangle_mesh(mesh_path, mesh)
 
-        f.write("\n")
 
-        for face in modes.fine_triangles:
-            f.write(f"{face[0]} {face[1]} {face[2]}\n")
+def write_splat_to_file(ply_file, output_folder):
+    if not os.path.exists(os.path.join(output_folder, ply_file)):
+        shutil.copyfile(ply_file, os.path.join(output_folder))
 
+
+def write_to_file(modes, output_folder, dataset_name, index, pcd, contact_point, direction, mesh, ply_file):
+    if not os.path.exists(dataset_name):
+        os.makedirs(dataset_name)
+
+    os.makedirs(os.path.join(dataset_name, output_folder), exist_ok=True)
+
+    # write pointcloud to file, only do this ONCE
+    write_pcd_to_file(dataset_name, output_folder, pcd)
+    write_splat_to_file(ply_file, output_folder)
+
+    # write the non_constants to file
+    write_segmentation_to_file(pcd, modes, mesh, dataset_name, output_folder, index)
+    write_impulse_to_file(contact_point, direction, dataset_name, output_folder, index)
+    write_mesh_to_file(dataset_name, output_folder, index, mesh)
